+++ app/src/main/java/org/hyperledger/besu/controller/QbftBesuControllerBuilder.java
@@ -283,7 +283,7 @@ public class QbftBesuControllerBuilder extends BesuControllerBuilder {
         new MessageTracker(qbftConfig.getDuplicateMessageLimit());
 
     final MessageFactory messageFactory =
-        new MessageFactory(nodeKey, blockEncoder, isLegacyRoundChangeEncodingEnabled);
+        new MessageFactory(nodeKey, blockEncoder, isLegacyBftProtocolEncodingEnabled);
 
     QbftRoundFactory qbftRoundFactory =
         new QbftRoundFactory(